requirements.txt
streamlit
torch
transformers
sentence-transformers
scikit-learn
numpy
nltk
faiss-cpu
rank-bm25
sentencepiece
protobuf
